Navbar, search, and iron:router
The navbar, right now, has an img in it with a src to /icon.svg. There
will need to be a setting for this image so that one can change the
image name and path to their liking.

The search is buggy. The Isotope animation is the only oddity.
Considering using Freewall to create the grid layout instead of Isotope.